remove sub ast mod
