+++ src/ast/mod.rs
@@ -1,3 +1,2 @@
-pub mod ast;
 pub mod dml;
 pub mod model;
