+++ src/grammar.lalrpop
@@ -1,5 +1,4 @@
 use crate::ast::{
-    ast::*,
     dml::*,
     model::*
 };
